pyproject.toml
[tool.poetry]
name = "ai-finance-assistant"
version = "1.0.0"
description = "Multi-agent AI system for financial education and portfolio management"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
# Core AI and LangChain
langchain = "^0.1.0"
langgraph = "^0.0.26"
openai = "^1.12.0"

# Vector Database
faiss-cpu = "^1.7.4"
sentence-transformers = "^2.2.2"

# Market Data
yfinance = "^0.2.12"
alpha-vantage = "^2.3.1"
requests = "^2.31.0"

# Web Interface
streamlit = "^1.31.0"
plotly = "^5.17.0"
matplotlib = "^3.8.2"
seaborn = "^0.13.0"

# Data Processing
pandas = "^2.1.4"
numpy = "^1.24.3"
scipy = "^1.11.4"

# Utilities
python-dotenv = "^1.0.0"
PyYAML = "^6.0.1"
pydantic = "^2.5.0"
typing-extensions = "^4.9.0"

[tool.poetry.group.dev.dependencies]
# Testing Framework
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
pytest-xdist = "^3.5.0"

# Code Quality
black = "^23.12.1"
isort = "^5.13.2"
flake8 = "^7.0.0"
mypy = "^1.8.0"

# Development Tools
jupyter = "^1.0.0"
ipython = "^8.18.1"
pre-commit = "^3.6.0"

# Documentation
sphinx = "^7.2.6"
sphinx-rtd-theme = "^2.0.0"

[tool.poetry.scripts]
finance-assistant = "src.web_app.main:main"
run-tests = "run_tests:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=html --cov-report=term"
markers = [
    "integration: marks tests as integration tests",
    "slow: marks tests as slow",
]
